fix: Notfound 페이지 구조에 맞게 변경
