+++ frontend/src/views/notfound/index.ts
@@ -1,0 +1,21 @@
+import View from '../../core/view';
+
+class NotfoundView extends View {
+  protected addListener (): void {
+    // no event
+  }
+
+  protected render (): void {
+    this.$target.innerHTML = `
+      <div class="notfound">
+        <h1>Notfound</h1>
+      </div>
+    `;
+  }
+
+  protected mount (): void {
+    // no mount
+  }
+}
+
+export default NotfoundView;
